requirements.txt


ffmpeg-python 0.2.0
future        0.18.2
MouseInfo     0.1.3
numpy         1.20.3
opencv-python 4.5.2.52
Pillow        8.2.0
pip           21.1.1
PyAutoGUI     0.9.52
PyGetWindow   0.0.9
PyMsgBox      1.0.9
pynput        1.7.3
pyperclip     1.8.2
PyRect        0.1.4
PyScreeze     0.1.27
PySimpleGUI   4.43.0
PyTweening    1.0.3
pywin32       301
setuptools    56.0.0
six           1.16.0
